FAC-963:impl:Move attributeinsight packages to insight package
