--- flickit-assessment-core/src/main/java/org/flickit/assessment/core/adapter/in/rest/attributeinsight/CreateAttributeAiInsightResponseDto.java
@@ -1,4 +1,0 @@
-package org.flickit.assessment.core.adapter.in.rest.attributeinsight;
-
-public record CreateAttributeAiInsightResponseDto(String content) {
-}
