+++ flickit-assessment-core/src/main/java/org/flickit/assessment/core/adapter/in/rest/insight/attributeinsight/CreateAttributeAiInsightResponseDto.java
@@ -1,0 +1,4 @@
+package org.flickit.assessment.core.adapter.in.rest.insight.attributeinsight;
+
+public record CreateAttributeAiInsightResponseDto(String content) {
+}
