+++ flickit-assessment-core/src/main/java/org/flickit/assessment/core/adapter/in/rest/insight/attributeinsight/GetAttributeInsightRestController.java
@@ -1,8 +1,8 @@
-package org.flickit.assessment.core.adapter.in.rest.attributeinsight;
+package org.flickit.assessment.core.adapter.in.rest.insight.attributeinsight;
 
 import lombok.RequiredArgsConstructor;
 import org.flickit.assessment.common.config.jwt.UserContext;
-import org.flickit.assessment.core.application.port.in.attributeinsight.GetAttributeInsightUseCase;
+import org.flickit.assessment.core.application.port.in.insight.attributeinsight.GetAttributeInsightUseCase;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
