+++ flickit-assessment-core/src/main/java/org/flickit/assessment/core/application/port/in/insight/attributeinsight/CreateAttributeAiInsightUseCase.java
@@ -1,4 +1,4 @@
-package org.flickit.assessment.core.application.port.in.attributeinsight;
+package org.flickit.assessment.core.application.port.in.insight.attributeinsight;
 
 import jakarta.validation.constraints.NotNull;
 import lombok.Builder;
