+++ flickit-assessment-core/src/test/java/org/flickit/assessment/core/application/service/insight/attributeinsight/GetAttributeInsightServiceTest.java
@@ -1,10 +1,10 @@
-package org.flickit.assessment.core.application.service.attributeinsight;
+package org.flickit.assessment.core.application.service.insight.attributeinsight;
 
 import org.flickit.assessment.common.application.domain.assessment.AssessmentAccessChecker;
 import org.flickit.assessment.common.exception.AccessDeniedException;
 import org.flickit.assessment.common.exception.ResourceNotFoundException;
 import org.flickit.assessment.core.application.domain.AssessmentResult;
-import org.flickit.assessment.core.application.port.in.attributeinsight.GetAttributeInsightUseCase;
+import org.flickit.assessment.core.application.port.in.insight.attributeinsight.GetAttributeInsightUseCase;
 import org.flickit.assessment.core.application.port.out.assessmentresult.LoadAssessmentResultPort;
 import org.flickit.assessment.core.application.port.out.attributeinsight.LoadAttributeInsightPort;
 import org.junit.jupiter.api.Test;
